@@ -1,39 +1,47 @@
 {
-  "name": "remote",
-  "version": "1.0.0",
-  "scripts": {
-    "build": "webpack --mode production",
-    "build:dev": "webpack --mode development",
-    "build:start": "cd dist && PORT=3002 npx serve",
-    "start": "webpack serve --open --mode development",
-    "start:live": "webpack serve --open --mode development --live-reload --hot"
-  },
-  "license": "MIT",
-  "author": {
-    "name": ""
-  },
-  "devDependencies": {
-    "@babel/core": "^7.15.8",
-    "@babel/plugin-transform-runtime": "^7.15.8",
-    "@babel/preset-env": "^7.15.8",
-    "@babel/preset-react": "^7.14.5",
-    "autoprefixer": "^10.1.0",
-    "babel-loader": "^8.2.2",
-    "css-loader": "^6.3.0",
-    "html-webpack-plugin": "^5.3.2",
-    "postcss": "^8.2.1",
-    "postcss-loader": "^4.1.0",
-    "style-loader": "^3.3.0",
-    "webpack": "^5.57.1",
-    "webpack-cli": "^4.9.0",
-    "webpack-dev-server": "^4.3.1"
-  },
-  "dependencies": {
-    "@mui/material": "^5.9.2",
-    "axios": "^0.27.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.3.0",
-    "weather-icons-react": "^1.2.0"
-  }
+	"name": "remote",
+	"version": "1.0.0",
+	"scripts": {
+		"start": "webpack serve --open --mode development --live-reload --hot",
+		"test": "jest"
+	},
+	"license": "MIT",
+	"author": {
+		"name": ""
+	},
+	"devDependencies": {
+		"@babel/core": "^7.15.8",
+		"@babel/plugin-transform-runtime": "^7.15.8",
+		"@babel/preset-env": "^7.15.8",
+		"@babel/preset-react": "^7.14.5",
+		"@testing-library/react": "^13.3.0",
+		"autoprefixer": "^10.1.0",
+		"babel-loader": "^8.2.2",
+		"css-loader": "^6.3.0",
+		"html-webpack-plugin": "^5.3.2",
+		"identity-obj-proxy": "^3.0.0",
+		"jest": "^28.1.3",
+		"jest-environment-jsdom": "^28.1.3",
+		"nock": "^13.2.9",
+		"postcss": "^8.2.1",
+		"postcss-loader": "^4.1.0",
+		"style-loader": "^3.3.0",
+		"webpack": "^5.57.1",
+		"webpack-cli": "^4.9.0",
+		"webpack-dev-server": "^4.3.1"
+	},
+	"dependencies": {
+		"@mui/material": "^5.9.2",
+		"axios": "^0.27.2",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-icons": "^4.4.0",
+		"react-router-dom": "^6.3.0"
+	},
+	"jest": {
+		"testEnvironment": "jsdom",
+		"moduleNameMapper": {
+			"\\.(css|less|scss|sass)$": "identity-obj-proxy"
+		}
+	}
 }
